put all tests together
